- Added support for automatically picking an invidious instance. (https://github.com/trizen/straw-viewer/issues/23)
Provided via:

	straw-viewer --api=auto

or by setting in the configuration file:

	api_host => "auto",

The list of instances ("instances.json") is cached locally and updated automatically every hour.

In order to distribute the load evenly across the invidious instances, when "api_host" is set to "auto", straw-viewer will pick a random instance every time the program is started.